new maskCurve code
It fits a tent-shaped function around each bin scaled to the bin's magnitude.

The old code did something similar,
but it didn't normalize the spread function fitted to each bin,
so the maskCurve had an upward tilt that had to be compensated by marginCurve.

I tweaked marginCurve to make then new code give roughly the same maskCurve,

The new code is easier to understand and mathamatically much prettier.
- The actually produces a ten shape in the log-log scale
- The amount of "spread" can now be offered as an adjustable parameter.
- It works consistently for different FFT sizes,
  but I have not found quality improvements with longer FFTs.

Performance is similar.