limit bin_gain slope and add new BIN_GAIN_DEBUG mode
Limit the slope of bin_gain to 6dB per octave (same as first order filter).
This got rid of most of the clipping distortion in the un-attenuated bins,
but it's still not always clean.
The allowed distortion level for attenuated needs to be done better.

Add a new debug mode that outputs the bin_gain'd but unclipped audio