Additional protection against additive distortion
This addresses the biggest problem with the integrated compressor,
which is the additive distortion (the kind that increases the level of a bin) popping up from time to time.
Now, in addition to the distortion limit relative to the unattenuated input,
there is an additional limit relative to the attenuated input that only applies to additive distortion.

TODO: Detect tone vs noise and allow more additive distortion in noise?
Not sure if tone detection can be done reliably with FFT256